refactor(worker): extract route-only asset helpers into routes/document-assets.ts
The previous shape had `tldraw-assets.ts` doing two unrelated jobs:
asset-storage / GC / lifecycle (long-running, called from
DocumentSyncRoom and from the document-delete handler) AND HTTP route
handling (multipart parsing, range-request math, content-type
derivation). Split them along usage:

- routes/document-assets.ts: chained Hono sub-router + every helper
  used only by the handlers (multipart parsing, content-type
  derivation, range-request math, the per-doc DB existence check, the
  asset-row insert, the post-upload GC scheduler). All file-private —
  the routes are their only consumers.
- tldraw-assets.ts: shrinks to GC / soft-delete lifecycle only
  (~363 lines, down from ~826). Exports unchanged for external
  consumers (DocumentSyncRoom imports `reconcileDocumentAssets`,
  `runDocumentAssetGc`, `isDocumentDeleting`,
  `getReferencedDocumentAssetNames`, `finalizeDeletingDocument`;
  `routes/documents.ts` imports `startDocumentDeletion`).

Cross-file leak is now exactly one helper: `getDocumentAssetKey` is
the only thing the routes side imports from tldraw-assets.ts. Both
sides need to derive the same R2 key (an upload PUT has to land where
a delete or GC sweep will look), so the function genuinely belongs in
one place. Documented inline.

Verified from cold: worker typecheck + 42 tests, app typecheck + 97
tests, full vite build, worker `build:types` emit — all green.
